Set background color on title slide
With the updated graphics for oss eu as the background image for the
title slide there is overlap between the graphics and the title text,
making the title page text partially illegible. This commit fixes this
by setting a static background color (chosen from the green gradient
used in the oss eu title slide image). While this is pretty ugly looking
it at least makes things legible. In the future I'll try to adjust the
bounding box size to improve the aesthetics, or come up with some other
technique to make the text legible with this background. This is just a
stop gap if I run out of time.