Brute force the title page to fit
This commit fixes the flawed approach taken in the previous commit to
make the text legible. Instead of trying to figure out a clever way to
make the built-in beamer titlepage template work this just builds a
custom template. It uses the tikzpicture package to manually layout all
the text on the titlepage so it doesn't overlap with any of the text
elements in the background image. It's not really a general purpose
solution because all the position is manually set.