Change color scheme
This commit changes the color scheme used for the presentation. Since
the title screen is green based (as opposed to the blue based used in
the OSS NA background slide) we should use green everywhere as the
secondary color in the slides. This commit takes a dark green from the
title page background image and uses that as the color where we were
using a blue before. This makes things more consistent throughout the
presentation.